test(javascript): address structured child review feedback
